refactor: 애그리거트 기준으로 도메인 패키지 구조 변경
- User와 Neighborhood 애그리거트를 기준으로 도메인 패키지 재구성
- 응집성과 애그리거트 간의 경계를 명확히 하기 위한 구조 개선
