+++ src/test/java/com/srltas/runtogether/domain/model/neighborhood/LocationUtilsTest.java
@@ -1,6 +1,6 @@
-package com.srltas.runtogether.domain.model.location;
+package com.srltas.runtogether.domain.model.neighborhood;
 
-import static com.srltas.runtogether.domain.model.location.LocationUtils.*;
+import static com.srltas.runtogether.domain.model.neighborhood.LocationUtils.*;
 import static org.hamcrest.MatcherAssert.*;
 import static org.hamcrest.Matchers.*;
